[E] Break up harvesting entity upsert action
Upserting harvest entities now fans out with a job for each entity
instead of a single job to handle all entities for a given record.

This handles situations where records that create many entities
(e.g. books with 200+ chapters) in order to prevent the job from
taking an unreasonable amount of time.