[F] Add frontend cache revalidation logic
Adds two new mutations:

* `frontendCacheRevalidateEntity`
* `frontendCacheRevalidateInstance`

Adds backgrounded revalidations to updateCommunity, updateCollection,
updateItem, and any time an entity's layouts are rerendered. The act
of revalidating an entity will also revalidate its parent in the
background, which will do the same for its parent, and so on all the
way up to the community. This ensures that if an entity is changed,
its changes are visible higher in the hierarchy.